App factory with db initialization
