+++ app/endpoints/search.py
@@ -1,10 +1,10 @@
-from fastapi import APIRouter, Depends, Request, Security
+from fastapi import APIRouter, Request, Security
 from fastapi.responses import JSONResponse
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.helpers._accesscontroller import AccessController
 from app.schemas.search import Searches, SearchRequest
-from app.sql.session import get_db as get_session
+from app.utils.depends import get_db_session
 from app.utils.context import global_context, request_context
 from app.utils.exceptions import CollectionNotFoundException
 from app.utils.variables import ENDPOINT__SEARCH
@@ -13,7 +13,7 @@ router = APIRouter()
 
 
 @router.post(path=ENDPOINT__SEARCH, dependencies=[Security(dependency=AccessController())], status_code=200, response_model=Searches)
-async def search(request: Request, body: SearchRequest, session: AsyncSession = Depends(get_session)) -> JSONResponse:
+async def search(request: Request, body: SearchRequest, session: AsyncSession = get_db_session()) -> JSONResponse:
     """
     Get relevant chunks from the collections and a query.
     """
