+++ app/tests/integ/test_identityaccessmanager.py
@@ -1,54 +1,52 @@
 import pytest
-from app.helpers._identityaccessmanager import IdentityAccessManager
-from app.schemas.auth import Limit, PermissionType
+from fastapi.testclient import TestClient
+from app.schemas.auth import PermissionType
+from app.utils.variables import ENDPOINT__ROLES
 
 
-@pytest.mark.usefixtures("client", "async_db_session")
+@pytest.mark.usefixtures("client")
 class TestIdentityAccessManager:
-    @pytest.mark.asyncio
-    async def test_update_role(self, async_db_session, client):
-        """Test the update_role function of IdentityAccessManager."""
-        async for session in async_db_session:
-            iam = IdentityAccessManager()
+    def test_update_role(self, client: TestClient):
+        """Test the update_role function through the API."""
 
-            # Create a role to update
-            role_name = "test-role"
-            limits = [
-                Limit(model="test-model", type="rpm", value=100),
-                Limit(model="test-model", type="rpd", value=1000),
-            ]
-            permissions = [PermissionType.CREATE_ROLE, PermissionType.CREATE_USER]
+        # Create a role to update
+        role_data = {
+            "name": "test-role",
+            "permissions": [PermissionType.CREATE_ROLE.value, PermissionType.CREATE_USER.value],
+            "limits": [
+                {"model": "test-model", "type": "rpm", "value": 100},
+                {"model": "test-model", "type": "rpd", "value": 1000},
+            ],
+        }
 
-            role_id = await iam.create_role(
-                session=session,
-                name=role_name,
-                limits=limits,
-                permissions=permissions,
-            )
+        # Create role via API
+        response = client.post_with_permissions(url=ENDPOINT__ROLES, json=role_data)
+        assert response.status_code == 201, response.text
+        role_id = response.json()["id"]
 
-            # Update the role
-            new_name = "updated-role"
-            new_limits = [
-                Limit(model="new-model", type="rpm", value=200),
-            ]
-            new_permissions = [PermissionType.DELETE_ROLE]
+        # Update the role
+        updated_role_data = {
+            "name": "updated-role",
+            "permissions": [PermissionType.DELETE_ROLE.value],
+            "limits": [
+                {"model": "new-model", "type": "rpm", "value": 200},
+            ],
+        }
 
-            await iam.update_role(
-                session=session,
-                role_id=role_id,
-                name=new_name,
-                limits=new_limits,
-                permissions=new_permissions,
-            )
+        # Update role via API
+        response = client.patch_with_permissions(url=f"{ENDPOINT__ROLES}/{role_id}", json=updated_role_data)
+        assert response.status_code == 204, response.text
 
-            # Fetch the updated role
-            roles = await iam.get_roles(session=session, role_id=role_id)
-            updated_role = roles[0]
+        # Fetch the updated role
+        response = client.get_with_permissions(url=f"{ENDPOINT__ROLES}/{role_id}")
+        assert response.status_code == 200, response.text
+        updated_role = response.json()
 
-            assert updated_role.name == new_name
-            assert len(updated_role.limits) == len(new_limits)
-            assert updated_role.limits[0].model == "new-model"
-            assert updated_role.limits[0].type == "rpm"
-            assert updated_role.limits[0].value == 200
-            assert len(updated_role.permissions) == len(new_permissions)
-            assert updated_role.permissions[0] == PermissionType.DELETE_ROLE
+        # Verify the updates
+        assert updated_role["name"] == "updated-role"
+        assert len(updated_role["limits"]) == 1
+        assert updated_role["limits"][0]["model"] == "new-model"
+        assert updated_role["limits"][0]["type"] == "rpm"
+        assert updated_role["limits"][0]["value"] == 200
+        assert len(updated_role["permissions"]) == 1
+        assert updated_role["permissions"][0] == PermissionType.DELETE_ROLE.value
