+++ app/tests/unit/test_accesscontroller.py
@@ -2,8 +2,6 @@ import pytest
 from unittest.mock import AsyncMock, MagicMock
 from fastapi import Request
 
-from app.helpers._accesscontroller import AccessController
-
 
 class MockUploadFile:
     """Mock UploadFile for testing"""
@@ -18,6 +16,8 @@ class MockUploadFile:
 class TestAccessController:
     @pytest.fixture
     def access_controller(self):
+        from app.helpers._accesscontroller import AccessController
+
         return AccessController()
 
     @pytest.mark.asyncio
